app/requirements.txt
# ===== Core Python deps =====
numpy<2

# ===== Web API =====
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-dotenv==1.0.1
python-multipart==0.0.9

# ===== Database =====
sqlalchemy==2.0.29
psycopg2-binary==2.9.9
pydantic==2.7.1

# ===== Embedding / Retrieval (RAG) =====
sentence-transformers==2.5.1
chromadb==0.5.0
pypdf==4.2.0

# ===== (Optional) Local LLM fallback =====
transformers==4.39.3
torch==2.2.0
accelerate==0.28.0
sentencepiece==0.1.99
safetensors==0.4.2

# ===== Gemini NEW SDK (รองรับ gemini-2.0-*) =====
google-genai>=1.0.0


